lint and biuld
